Bump vyper version for CurveCryptoSwap2 0.3.1 -> 0.3.3

+++ contracts/two/CurveCryptoSwap2.vy
@@ -1,4 +1,4 @@
-# @version 0.3.1
+# @version 0.3.3
 # (c) Curve.Fi, 2021
 # Pool for two crypto assets
 
@@ -87,7 +87,7 @@ event ClaimAdminFee:
     tokens: uint256
 
 
-N_COINS: constant(int128) = 2
+N_COINS: constant(uint256) = 2
 PRECISION: constant(uint256) = 10 ** 18  # The precision to convert to
 A_MULTIPLIER: constant(uint256) = 10000
